refactor to not have duplicate code
