+++ lib/base-compiler.js
@@ -36,7 +36,8 @@ const temp = require('temp'),
     argumentParsers = require("./compilers/argument-parsers"),
     CompilerArguments = require("./compiler-arguments"),
     cfg = require('./cfg'),
-    languages = require('./languages').list;
+    languages = require('./languages').list,
+    toolchainUtils = require('./toolchain-utils');
 
 class BaseCompiler {
     constructor(compilerInfo, env) {
@@ -64,7 +65,7 @@ class BaseCompiler {
         this.asm = new AsmParser(this.compilerProps);
         this.llvmIr = new LlvmIrParser(this.compilerProps);
 
-        this.toolchainPath = this.getToolchainPath();
+        this.toolchainPath = toolchainUtils.getToolchainPath(this.compiler.exe, this.compiler.options);
 
         this.possibleArguments = new CompilerArguments(this.compiler.id);
         this.possibleTools = _.values(compilerInfo.tools);
@@ -210,24 +211,6 @@ class BaseCompiler {
         return _.find(foundLib.versions, (o, versionId) => versionId === selectedLib.version);
     }
 
-    getToolchainPath() {
-        const options = this.compiler.options ? this.compiler.options.split(" ") : [];
-        const existingChain = _.find(options, (elem) => elem.indexOf("--gcc-toolchain=") !== -1);
-
-        if (!existingChain) {
-            const gxxname = _.find(options, (elem) => elem.indexOf("--gxx-name=") !== -1);
-            if (gxxname) {
-                return path.resolve(path.dirname(gxxname.substr(11)), "..");
-            } else if (this.compiler.exe && this.compiler.exe.indexOf("/g++") !== -1) {
-                return path.resolve(path.dirname(this.compiler.exe), "..");
-            } else {
-                return false;
-            }
-        } else {
-            return existingChain.substr(16);
-        }
-    }
-
     findAutodetectStaticLibLink(linkname) {
         const foundLib = _.findKey(this.compiler.libs, (lib) => {
             return (lib.versions.autodetect && lib.versions.autodetect.staticliblink &&
@@ -356,7 +339,7 @@ class BaseCompiler {
         if (this.toolchainPath) {
             toolchainLibraryPaths = [
                 path.join(this.toolchainPath, '/lib64'),
-                path.join(this.toolchainPath, '/lib')
+                path.join(this.toolchainPath, '/lib32')
             ];
         }
