+++ lib/toolchain-utils.js
@@ -1,0 +1,49 @@
+// Copyright (c) 2020, Compiler Explorer Authors
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright notice,
+//       this list of conditions and the following disclaimer.
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+"use strict";
+
+const _ = require('underscore'),
+    path = require('path');
+
+function getToolchainPath(compilerExe, compilerOptions) {
+    const options = compilerOptions ? compilerOptions.split(" ") : [];
+    const existingChain = _.find(options, (elem) => elem.indexOf("--gcc-toolchain=") !== -1);
+
+    if (!existingChain) {
+        const gxxname = _.find(options, (elem) => elem.indexOf("--gxx-name=") !== -1);
+        if (gxxname) {
+            return path.resolve(path.dirname(gxxname.substr(11)), "..");
+        } else if (typeof compilerExe === 'string' && compilerExe.indexOf("/g++") !== -1) {
+            return path.resolve(path.dirname(compilerExe), "..");
+        } else {
+            return false;
+        }
+    } else {
+        return existingChain.substr(16);
+    }
+}
+
+module.exports = {
+    getToolchainPath: getToolchainPath
+};
